Modularize chrome/browser/ui/views/translate/
Move the //chrome/browser/ui/views/translate sources out of the
monolithic //chrome/browser/ui:ui target into a dedicated
//chrome/browser/ui/views/translate GN target, following the established
modularization pattern (e.g. //chrome/browser/ui/views/commerce).

New chrome/browser/ui/views/translate/BUILD.gn defines:
- :translate    - the public headers and their public_deps.
- :impl         - the .cc files. The remaining circular includes back
                  into //chrome/browser/ui:ui (chrome_pages.h,
                  scoped_tabbed_browser_displayer.h,
                  views/frame/browser_view.h) are handled via
                  browser_public_dependencies plus a documented entry in
                  the ui target's allow_circular_includes_from list.
- :test_support - translate_bubble_test_utils_views.cc, previously
                  duplicated inline across four test targets.
- :unit_tests, :browser_tests, :interactive_ui_tests - the test sources,
                  moved out of //chrome/test.

Also fix a latent -Wsign-compare error in
translate_bubble_view_browsertest.cc (int vs size_t in EXPECT_EQ),
surfaced by recompiling the file in the new target.

R=chromium-bedrock-reviews@google.com

Bug: 353332589
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/8003408
Commit-Queue: Antonio Gomes Netto <tonikitoo@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1652676}

+++ chrome/browser/ui/views/translate/translate_bubble_view_browsertest.cc
@@ -318,7 +318,8 @@ IN_PROC_BROWSER_TEST_F(TranslateBubbleViewBrowserTest_SearchUIDisabled,
                                                            u"Spanish");
 
   EXPECT_EQ(bubble->model()->GetTargetLanguageIndex(),
-            bubble->model()->GetTargetLanguageIndexForCode("es").value());
+            static_cast<int>(
+                bubble->model()->GetTargetLanguageIndexForCode("es").value()));
 }
 
 IN_PROC_BROWSER_TEST_F(TranslateBubbleViewBrowserTest_SearchUIEnabled,
@@ -339,7 +340,8 @@ IN_PROC_BROWSER_TEST_F(TranslateBubbleViewBrowserTest_SearchUIEnabled,
                                                            u"Spanish");
 
   EXPECT_EQ(bubble->model()->GetTargetLanguageIndex(),
-            bubble->model()->GetTargetLanguageIndexForCode("es").value());
+            static_cast<int>(
+                bubble->model()->GetTargetLanguageIndexForCode("es").value()));
 }
 
 IN_PROC_BROWSER_TEST_F(TranslateBubbleViewBrowserTest_SearchUIEnabled,
